# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TestAtlas API",
    "version": "baseline"
  },
  "paths": {
...
    "/api/resolve": {
      "post": {
        "tags": [
          "resolve"
        ],
        "summary": "Resolve",
        "description": "이름으로 계열·기종·기준정보 값·시험법을 찾는다.\n\n**무엇을 만들기 전에 이것을 먼저 부른다.** 응답의 `match` 가 셋이다:\n\n    exact       하나로 정해졌다. `id` 를 그대로 쓴다\n    candidates  여럿이다. 고르지 말고 사람에게 묻는다\n    none        없다. 새로 만들거나 비워 둔다 — **지어내지 않는다**\n\n못 찾은 것은 실패가 아니라서 언제나 200 이다. `hint` 에 다음에 할 일이 한 줄로\n적혀 있다.",
        "operationId": "resolve_api_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/equipment": {
      "get": {
        "tags": [
          "equipment"
        ],
        "summary": "List Equipment",
        "operationId": "list_equipment_api_equipment_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "workspace",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EquipmentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "equipment"
        ],
        "summary": "Create Equipment",
        "operationId": "create_equipment_api_equipment_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/equipment/{equipment_id}": {
      "get": {
        "tags": [
          "equipment"
        ],
        "summary": "Read Equipment",
        "operationId": "read_equipment_api_equipment__equipment_id__get",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Equipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "equipment"
        ],
        "summary": "Update Equipment",
        "operationId": "update_equipment_api_equipment__equipment_id__patch",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Equipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "equipment"
        ],
        "summary": "Delete Equipment",
        "operationId": "delete_equipment_api_equipment__equipment_id__delete",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Equipment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/equipment/{equipment_id}/calibrations": {
      "get": {
        "tags": [
          "equipment"
        ],
        "summary": "List Calibrations",
        "operationId": "list_calibrations_api_equipment__equipment_id__calibrations_get",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Equipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalibrationOut"
                  },
                  "title": "Response List Calibrations Api Equipment  Equipment Id  Calibrations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "equipment"
        ],
        "summary": "Add Calibration",
        "operationId": "add_calibration_api_equipment__equipment_id__calibrations_post",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Equipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/equipment-series": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Series",
        "operationId": "list_series_api_equipment_series_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "category_term_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Term Id"
            }
          },
          {
            "name": "owned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Owned"
            }
          },
          {
            "name": "issue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(capabilities)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EquipmentSeriesOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create Series",
        "description": "계열을 만든다.\n\n**먼저 `POST /api/resolve` 로 찾는다.** 같은 계열이 두 줄로 갈리면 보유 장비가\n어느 쪽을 가리켰는지에 따라 검색 결과가 나뉜다.\n\n이미 있으면 409 와 함께 `details.series_id` 가 온다 — **409 는 실패가 아니라\n답이다.** 그 id 를 그대로 쓰면 된다.\n\n제조사·분류는 id 대신 이름(`maker`·`category`)으로 줘도 된다. 하나로 정해지지\n않으면 400 과 후보 목록이 온다 — **고르지 말고 사람에게 물어라.**",
        "operationId": "create_series_api_equipment_series_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentSeriesCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentSeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/equipment-series/{series_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Read Series",
        "operationId": "read_series_api_equipment_series__series_id__get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentSeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Update Series",
        "operationId": "update_series_api_equipment_series__series_id__patch",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentSeriesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentSeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "catalog"
        ],
        "summary": "Delete Series",
        "operationId": "delete_series_api_equipment_series__series_id__delete",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResolveRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "pattern": "^(series|model|term|method)$",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Text"
          },
          "axis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Axis"
          },
          "maker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maker"
          },
          "limit": {
            "type": "integer",
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "kind",
          "text"
        ],
        "title": "ResolveRequest",
        "description": "「이게 이미 있나?」\n\nAI 가 하는 첫 동작은 언제나 이것이다. 목록 검색(`?q=`)으로 흉내낼 수는 있지만,\n**흉내는 매번 조금씩 다르게 틀린다** — 「확실히 이것」 과 「후보 셋」 과 「없음」 을\n구별해 주지 않기 때문이다."
      },
      "ResolveResponse": {
        "properties": {
          "match": {
            "type": "string",
            "title": "Match"
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ResolveCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "hint": {
            "type": "string",
            "title": "Hint"
          }
        },
        "type": "object",
        "required": [
          "match",
          "id",
          "label",
          "candidates",
          "hint"
        ],
        "title": "ResolveResponse",
        "description": "찾은 결과.\n\n`match` 가 세 값인 것이 요점이다.\n\n    exact       하나로 정해졌다. `id` 를 그대로 쓰면 된다\n    candidates  여럿이다. **고르지 말고 사람에게 물어라**\n    none        없다. 만들거나, 비워 두거나 — 지어내지는 마라"
      },
...
    }
  }
}